requirements.txt
django==2.2.8
requests==2.21.0
djangorestframework==3.8.2
django-cors-headers==3.2.1
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
psycopg2==2.8.5
# mongoengine==0.19.1
# django-rest-framework-mongoengine==3.4.1
gunicorn==20.0.4
# celery==4.3.0
cerberus==1.3.2
# channels==2.4.0
